audio_scripts/requirements.txt
deprecation==2.1.0
eyeD3==0.9.8
filetype==1.2.0
greenlet==3.2.3
msgpack==1.1.1
packaging==25.0
pynvim==0.5.2
python-slugify==8.0.4
text-unidecode==1.3
titlecase==2.4.1
typing_extensions==4.14.1
